+++ src/7.x.ts
@@ -46,20 +46,22 @@ const getAccountAuthorization = async ({
   provider,
   email,
   password,
+  fetch: customFetch,
 }: {
   provider: string
   email: string
   password: string
+  fetch: typeof globalThis.fetch
 }) => {
   // First we request the account API controls to find out where we can log in
-  const indexResponse = await fetch(new URL('.account/', provider))
+  const indexResponse = await customFetch(new URL('.account/', provider))
 
   await throwIfResponseNotOk(indexResponse)
 
   const { controls } = (await indexResponse.json()) as AccountHandles
 
   // And then we log in to the account API
-  const response = await fetch(controls.password.login, {
+  const response = await customFetch(controls.password.login, {
     method: 'POST',
     headers: { 'content-type': 'application/json' },
     body: JSON.stringify({ email, password }),
@@ -76,14 +78,16 @@ const getIdAndSecret = async ({
   authorization,
   provider,
   webId,
+  fetch: customFetch,
 }: {
   authorization: string
   provider: string
   webId?: string
+  fetch: typeof globalThis.fetch
 }) => {
   // Now that we are logged in, we need to request the updated controls from the server.
   // These will now have more values than in the previous example.
-  const indexResponse = await fetch(new URL('.account/', provider), {
+  const indexResponse = await customFetch(new URL('.account/', provider), {
     headers: { authorization: `CSS-Account-Token ${authorization}` },
   })
 
@@ -92,24 +96,28 @@ const getIdAndSecret = async ({
   const handles = (await indexResponse.json()) as AccountHandles
 
   // Here we request the server to generate a token on our account
-  const response = await fetch(handles.controls.account.clientCredentials, {
-    method: 'POST',
-    headers: {
-      authorization: `CSS-Account-Token ${authorization}`,
-      'content-type': 'application/json',
-    },
-    // The name field will be used when generating the ID of your token.
-    // The WebID field determines which WebID you will identify as when using the token.
-    // Only WebIDs linked to your account can be used.
-    body: JSON.stringify({
-      name: 'my-token',
-      webId: await getWebId({
-        webId,
-        handles,
-        authorization,
+  const response = await customFetch(
+    handles.controls.account.clientCredentials,
+    {
+      method: 'POST',
+      headers: {
+        authorization: `CSS-Account-Token ${authorization}`,
+        'content-type': 'application/json',
+      },
+      // The name field will be used when generating the ID of your token.
+      // The WebID field determines which WebID you will identify as when using the token.
+      // Only WebIDs linked to your account can be used.
+      body: JSON.stringify({
+        name: 'my-token',
+        webId: await getWebId({
+          webId,
+          handles,
+          authorization,
+          fetch: customFetch,
+        }),
       }),
-    }),
-  })
+    },
+  )
 
   await throwIfResponseNotOk(response)
 
@@ -129,14 +137,16 @@ const getWebId = async ({
   webId,
   handles,
   authorization,
+  fetch: customFetch,
 }: {
   webId?: string
   handles: AccountHandles
   authorization: string
+  fetch: typeof globalThis.fetch
 }) => {
   // now let's get all available webIds
 
-  const webIdsResponse = await fetch(handles.controls.account.webId, {
+  const webIdsResponse = await customFetch(handles.controls.account.webId, {
     headers: {
       authorization: `CSS-Account-Token ${authorization}`,
     },
@@ -165,10 +175,12 @@ const getAccessToken = async ({
   id,
   secret,
   provider,
+  fetch: customFetch,
 }: {
   id: string
   secret: string
   provider: string
+  fetch: typeof globalThis.fetch
 }) => {
   // A key pair is needed for encryption.
   // This function from `solid-client-authn` generates such a pair for you.
@@ -181,7 +193,7 @@ const getAccessToken = async ({
   // http://localhost:3000/.well-known/openid-configuration
   // if your server is hosted at http://localhost:3000/.
   const tokenUrl = new URL('.oidc/token', provider)
-  const response = await fetch(tokenUrl, {
+  const response = await customFetch(tokenUrl, {
     method: 'POST',
     headers: {
       // The header needs to be in base64 encoding.
@@ -207,13 +219,18 @@ const getAccessToken = async ({
 const generateAuthenticatedFetch = async ({
   accessToken,
   dpopKey,
+  fetch: customFetch,
 }: {
   accessToken: string
   dpopKey: KeyPair
+  fetch: typeof globalThis.fetch
 }) => {
   // The DPoP key needs to be the same key as the one used in the previous step.
   // The Access token is the one generated in the previous step.
-  const authFetch = buildAuthenticatedFetch(accessToken, { dpopKey })
+  const authFetch = buildAuthenticatedFetch(accessToken, {
+    dpopKey,
+    fetch: customFetch,
+  })
 
   return authFetch
 }
@@ -223,29 +240,38 @@ export const getAuthenticatedFetch = async ({
   email,
   password,
   webId,
+  fetch: customFetch = globalThis.fetch,
 }: {
   provider: string
   email: string
   password: string
   webId?: string
+  fetch?: typeof globalThis.fetch
 }) => {
   const authorization = await getAccountAuthorization({
     provider,
     email,
     password,
+    fetch: customFetch,
   })
 
   const { id, secret } = await getIdAndSecret({
     authorization,
     provider,
     webId,
+    fetch: customFetch,
   })
   const { accessToken, dpopKey } = await getAccessToken({
     id,
     secret,
     provider,
+    fetch: customFetch,
   })
-  return await generateAuthenticatedFetch({ accessToken, dpopKey })
+  return await generateAuthenticatedFetch({
+    accessToken,
+    dpopKey,
+    fetch: customFetch,
+  })
 }
 
 const throwIfResponseNotOk = async (response: Response) => {
@@ -262,11 +288,13 @@ export const createAccount = async ({
   password,
   email,
   provider,
+  fetch: customFetch = globalThis.fetch,
 }: {
   username: string
   password: string
   email: string
   provider: string
+  fetch?: typeof globalThis.fetch
 }) => {
   const config = {
     idp: new URL('./', provider).toString(),
@@ -280,28 +308,31 @@ export const createAccount = async ({
   const accountEndpoint = new URL('.account/account/', provider)
 
   // create the account
-  const response = await fetch(accountEndpoint, { method: 'POST' })
+  const response = await customFetch(accountEndpoint, { method: 'POST' })
   await throwIfResponseNotOk(response)
 
   const account = (await response.json()) as { authorization: string }
   const authorization = `CSS-Account-Token ${account.authorization}`
 
   // get account handles
-  const response2 = await fetch(new URL('.account/', provider), {
+  const response2 = await customFetch(new URL('.account/', provider), {
     headers: { authorization },
   })
   await throwIfResponseNotOk(response2)
 
   const handles = (await response2.json()) as AccountHandles
 
-  const createLoginResponse = await fetch(handles.controls.password.create, {
-    method: 'POST',
-    body: JSON.stringify({ email, password, confirmPassword: password }),
-    headers: { 'content-type': 'application/json', authorization },
-  })
+  const createLoginResponse = await customFetch(
+    handles.controls.password.create,
+    {
+      method: 'POST',
+      body: JSON.stringify({ email, password, confirmPassword: password }),
+      headers: { 'content-type': 'application/json', authorization },
+    },
+  )
   await throwIfResponseNotOk(createLoginResponse)
 
-  const response3 = await fetch(handles.controls.account.pod, {
+  const response3 = await customFetch(handles.controls.account.pod, {
     method: 'POST',
     headers: { 'content-type': 'application/json', authorization },
     body: JSON.stringify({ name: username }),