+++ src/tests/v7.getAuthenticatedFetch.spec.ts
@@ -1,13 +1,23 @@
 import css from '@solid/community-server'
 import assert from 'node:assert'
 import { randomUUID } from 'node:crypto'
-import { after, before, describe, it } from 'node:test'
+import {
+  after,
+  afterEach,
+  before,
+  beforeEach,
+  describe,
+  it,
+  Mock,
+  mock,
+} from 'node:test'
 import { v7 } from '../index.js'
 import { getRandomPort, startServer, stopServer } from './setup.js'
 
 describe('v7 getAuthenticatedFetch', async () => {
   let port = -1
   let server: css.App
+  let fetchSpy: Mock<typeof globalThis.fetch>
 
   before(async () => {
     port = getRandomPort()
@@ -16,6 +26,14 @@ describe('v7 getAuthenticatedFetch', async () => {
 
   after(async () => {
     await stopServer(server)
+  })
+
+  beforeEach(() => {
+    fetchSpy = mock.fn(globalThis.fetch)
+  })
+
+  afterEach(() => {
+    fetchSpy.mock.restore()
   })
 
   const options = [
@@ -46,10 +64,36 @@ describe('v7 getAuthenticatedFetch', async () => {
         },
       },
     },
+    {
+      condition: 'with custom fetch',
+      params: {
+        get provider() {
+          const url = new URL('http://localhost')
+          url.port = String(port)
+          return url.toString()
+        },
+        email: 'person@example',
+        password: 'correcthorsebatterystaple',
+        get fetch() {
+          return fetchSpy
+        },
+      },
+      assert: {
+        prepared: () => {
+          assert.ok(fetchSpy.mock.callCount() > 0)
+          fetchSpy.mock.resetCalls()
+        },
+        finished: () => {
+          assert.equal(fetchSpy.mock.callCount(), 2)
+        },
+      },
+    },
   ]
   options.forEach(options => {
     it(`[${options.condition}] should produce functioning authenticated fetch`, async () => {
       const authFetch = await v7.getAuthenticatedFetch(options.params)
+
+      options.assert?.prepared()
 
       const filename = randomUUID() + '.ttl'
       const testFile = new URL(filename, options.params.provider)
@@ -72,6 +116,8 @@ describe('v7 getAuthenticatedFetch', async () => {
           '@type': [testFile + '#test'],
         },
       ])
+
+      options.assert?.finished()
     })
   })
 })